pyproject.toml
[tool.poetry]
name = "mytgbot"
version = "0.1.0"
description = ""
authors = ["alim_b <alim.bayr@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.8"
pyTelegramBotAPI = "^4.10.0"
python-dotenv = "^0.19.2"
peewee = "^3.15.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
